+++ MapBrowser/Core/Configurator.cs
@@ -11,6 +11,7 @@ sealed class Configurator : IConfigurator {
   public void Configure(IContainerDefinition containerDefinition) {
     containerDefinition.Bind<WorkshopMetadataService>().AsSingleton();
     containerDefinition.Bind<WorkshopLiveDetailsService>().AsSingleton();
+    containerDefinition.Bind<WorkshopMapSizeService>().AsSingleton();
     containerDefinition.Bind<WorkshopSubscriptionService>().AsSingleton();
   }
 }